1540 - WIP - table header tooltips - style of tooltip is now ok
